# Remove unnecessary assignment to fix type instability in ForwardDiff.hessian!