Before:
```julia
julia> using Traceur

julia> @trace ForwardDiff.hessian!(hr2, f2, max_density_params, cfg2)
(ForwardDiff.hessian!)(::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}, ::#f2, ::Array{Float64,1}, ::ForwardDiff.HessianConfig{ForwardDiff.Tag{#f2,Float64},Float64,8,Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},8},1},Tuple{Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1},Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1}}}, ::Val{true}) at /home/chris/.julia/v0.6/ForwardDiff/src/hessian.jl:47
  result is assigned as DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}} at line 47
  result is assigned as Core.Box at line 47
  dynamic dispatch to ($(QuoteNode(DiffResults.hessian)))((Core.getfield)(_9::Core.Box, :contents)) at line 54
  dynamic dispatch to ($(QuoteNode(DiffResults.gradient)))((Core.getfield)(_9::Core.Box, :contents)) at line 54
  returns Any
(ForwardDiff.hessian!)(::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}, ::#f2, ::Array{Float64,1}, ::ForwardDiff.HessianConfig{ForwardDiff.Tag{#f2,Float64},Float64,8,Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},8},1},Tuple{Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1},Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1}}}) at /home/chris/.julia/v0.6/ForwardDiff/src/hessian.jl:47
  returns Any
MutableDiffResult(14774.342219725911, ([1.50629e-8, 2.16231e-7, -1.52464e-7, -1.84819e-8, -4.70467e-8, -1.71915e-7, 1.96066e-7, -1.81974e-7], [86.4397 1.3505 -86.697 42.0846 18.541 41.7884 60.8334 80.3928; 1.3505 11.59 38.7777 -5.84322 8.49931 13.6476 20.3259 26.363; -86.697 38.7777 734.801 -4.81614 -134.535 -161.619 -167.768 -166.47; 42.0846 -5.84322 -4.81614 54.096 -17.7162 -23.4584 -21.8144 -20.9364; 18.541 8.49931 -134.535 -17.7162 256.49 -0.0 -0.0 -0.0; 41.7884 13.6476 -161.619 -23.4584 -0.0 502.43 -0.0 -0.0; 60.8334 20.3259 -167.768 -21.8144 -0.0 -0.0 694.565 -0.0; 80.3928 26.363 -166.47 -20.9364 -0.0 -0.0 -0.0 825.356]))


julia> @code_warntype ForwardDiff.hessian!(hr2, f2, max_density_params, cfg2)
Variables:
  #self#::ForwardDiff.#hessian!
  result::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}
  f::#f2
  x::Array{Float64,1}
  cfg::ForwardDiff.HessianConfig{ForwardDiff.Tag{#f2,Float64},Float64,8,Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},8},1},Tuple{Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1},Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1}}}

Body:
  begin 
      return $(Expr(:invoke, MethodInstance for hessian!(::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}, ::Function, ::Array{Float64,1}, ::ForwardDiff.HessianConfig{ForwardDiff.Tag{#f2,Float64},Float64,8,Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},8},1},Tuple{Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1},Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1}}}, ::Val{true}), :(#self#), :(result), :(f), :(x), :(cfg), :($(QuoteNode(Val{true}())))))
  end::Any
```

After:
```julia
julia> using Traceur

julia> @trace ForwardDiff.hessian!(hr2, f2, max_density_params, cfg2)
MutableDiffResult(14774.342219725911, ([1.50629e-8, 2.16231e-7, -1.52464e-7, -1.84819e-8, -4.70467e-8, -1.71915e-7, 1.96066e-7, -1.81974e-7], [86.4397 1.3505 -86.697 42.0846 18.541 41.7884 60.8334 80.3928; 1.3505 11.59 38.7777 -5.84322 8.49931 13.6476 20.3259 26.363; -86.697 38.7777 734.801 -4.81614 -134.535 -161.619 -167.768 -166.47; 42.0846 -5.84322 -4.81614 54.096 -17.7162 -23.4584 -21.8144 -20.9364; 18.541 8.49931 -134.535 -17.7162 256.49 -0.0 -0.0 -0.0; 41.7884 13.6476 -161.619 -23.4584 -0.0 502.43 -0.0 -0.0; 60.8334 20.3259 -167.768 -21.8144 -0.0 -0.0 694.565 -0.0; 80.3928 26.363 -166.47 -20.9364 -0.0 -0.0 -0.0 825.356]))

julia> @code_warntype ForwardDiff.hessian!(hr2, f2, max_density_params, cfg2)
Variables:
  #self#::ForwardDiff.#hessian!
  result::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}
  f::#f2
  x::Array{Float64,1}
  cfg::ForwardDiff.HessianConfig{ForwardDiff.Tag{#f2,Float64},Float64,8,Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},8},1},Tuple{Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1},Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1}}}

Body:
  begin 
      return $(Expr(:invoke, MethodInstance for hessian!(::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}, ::Function, ::Array{Float64,1}, ::ForwardDiff.HessianConfig{ForwardDiff.Tag{#f2,Float64},Float64,8,Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},8},1},Tuple{Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1},Array{ForwardDiff.Dual{ForwardDiff.Tag{#f2,Float64},Float64,8},1}}}, ::Val{true}), :(#self#), :(result), :(f), :(x), :(cfg), :($(QuoteNode(Val{true}())))))
  end::DiffResults.MutableDiffResult{2,Float64,Tuple{Array{Float64,1},Array{Float64,2}}}
```

The actual performance difference is negligible, although if the type instabilities "infected" other parts of your code, it could be a problem. Before:
```julia
julia> @benchmark ForwardDiff.hessian!($hr2, f2, $max_density_params, $cfg2)
BenchmarkTools.Trial: 
  memory estimate:  6.08 KiB
  allocs estimate:  14
  --------------
  minimum time:     17.984 μs (0.00% GC)
  median time:      18.906 μs (0.00% GC)
  mean time:        19.678 μs (1.30% GC)
  maximum time:     1.341 ms (96.11% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
after
```julia
julia> @benchmark ForwardDiff.hessian!($hr2, f2, $max_density_params, $cfg2)
BenchmarkTools.Trial: 
  memory estimate:  6.05 KiB
  allocs estimate:  12
  --------------
  minimum time:     18.235 μs (0.00% GC)
  median time:      19.136 μs (0.00% GC)
  mean time:        19.949 μs (1.60% GC)
  maximum time:     1.660 ms (96.77% GC)
  --------------
  samples:          10000
  evals/sample:     1
```
These benchmarks were prior to merging the commits for specializing gradients on input functions, after which:
```julia
julia> @benchmark ForwardDiff.hessian!($hr2, f2, $max_density_params, $cfg2)
BenchmarkTools.Trial: 
  memory estimate:  5.39 KiB
  allocs estimate:  11
  --------------
  minimum time:     17.653 μs (0.00% GC)
  median time:      20.097 μs (0.00% GC)
  mean time:        20.798 μs (1.66% GC)
  maximum time:     1.791 ms (96.75% GC)
  --------------
  samples:          10000
  evals/sample:     1
```