@KristofferC 2018-04-11
https://github.com/JuliaDiff/ForwardDiff.jl/blob/4a621653fc717f8501df29ecba2854e657f43ba1/src/gradient.jl#L69-L73

same thing?